+++ src/clj_cbor/tags/clojure.clj
@@ -1,19 +1,5 @@
 (ns clj-cbor.tags.clojure
-  "Read and write handler support for Clojure types.
-
-  Keywords and symbols are represented using tag 39 ('identifier') applied to
-  the string version of the value. This adds three bytes to the size of the
-  identifier itself for the header, tag code, and string header. Keywords are
-  symbols whose first character is a colon (:).
-
-  Tagged literals are represented using tag 27 ('generic object') applied to an
-  array containing two elements. The first element is the string version of the
-  EDN tag symbol and the second is the tagged literal form.
-
-  See:
-  - https://github.com/lucas-clemente/cbor-specs/blob/master/id.md
-  - http://cbor.schmorp.de/generic-object
-  "
+  "Read and write handler support for Clojure types."
   (:require
     [clj-cbor.data.core :as data])
   (:import
@@ -25,17 +11,27 @@
 
 ;; ## Symbols & Keywords
 
+(def ^:const identifier-tag
+  "Keywords and symbols are represented using tag 39 ('identifier') applied to
+  the string version of the value. This adds three bytes to the size of the
+  identifier itself for the header, tag code, and string header. Keywords are
+  symbols whose first character is a colon (:).
+
+  See: [https://github.com/lucas-clemente/cbor-specs/blob/master/id.md](https://github.com/lucas-clemente/cbor-specs/blob/master/id.md)"
+  39)
+
+
 (defn format-symbol
   [value]
-  (data/tagged-value 39 (str value)))
+  (data/tagged-value identifier-tag (str value)))
 
 
 (defn parse-symbol
-  [tag value]
+  [value]
   (when-not (string? value)
     (throw (ex-info (str "Symbols must be tagged strings, got: "
                          (class value))
-                    {:tag tag, :value value})))
+                    {:value value})))
   (if (= \: (first value))
     (keyword (subs value 1))
     (symbol value)))
@@ -44,20 +40,28 @@
 
 ;; ## Tagged Literals
 
-;; Tag 27
-;; http://cbor.schmorp.de/generic-object
+(def ^:const generic-object-tag
+  "Tagged literals are represented using tag 27 ('generic object') applied to
+  an array containing two elements. The first element is the string version of
+  the EDN tag symbol and the second is the tagged literal form.
+
+  See: [http://cbor.schmorp.de/generic-object](http://cbor.schmorp.de/generic-object)"
+  27)
+
 
 (defn format-tagged-literal
   [value]
-  (data/tagged-value 27 [(str (:tag value)) (:form value)]))
+  (data/tagged-value
+    generic-object-tag
+    [(str (:tag value)) (:form value)]))
 
 
 (defn parse-tagged-literal
-  [tag value]
+  [value]
   (when-not (and (sequential? value) (= 2 (count value)))
     (throw (ex-info (str "Sets must be tagged two-element arrays, got: "
                          (class value))
-                    {:tag tag, :value value})))
+                    {:value value})))
   (tagged-literal (symbol (first value)) (second value)))
 
 
@@ -73,5 +77,5 @@
 
 (def clojure-read-handlers
   "Map of tag codes to read handlers to parse Clojure values."
-  {27 parse-tagged-literal
-   39 parse-symbol})
+  {generic-object-tag parse-tagged-literal
+   identifier-tag     parse-symbol})