+++ test/clj_cbor/tags/text_test.clj
@@ -12,7 +12,7 @@
 (deftest uri-coding
   (testing "parsing checks"
     (is (thrown-with-msg? Exception #"must be tagged strings"
-          (parse-uri 32 (byte-array 4)))))
+          (parse-uri (byte-array 4)))))
   (with-codec {:write-handlers text-write-handlers
                :read-handlers text-read-handlers}
     (check-roundtrip (URI. "http://www.example.com") "D82076687474703A2F2F7777772E6578616D706C652E636F6D")))
@@ -21,7 +21,7 @@
 (deftest pattern-coding
   (testing "parsing checks"
     (is (thrown-with-msg? Exception #"must be tagged strings"
-          (parse-pattern 35 (byte-array 4)))))
+          (parse-pattern (byte-array 4)))))
   (with-codec {:write-handlers text-write-handlers
                :read-handlers text-read-handlers}
     (check-roundtrip str #"abc123" "D82366616263313233")))
@@ -30,7 +30,7 @@
 (deftest uuid-coding
   (testing "parsing checks"
     (is (thrown-with-msg? Exception #"must be tagged byte strings"
-          (parse-uuid 37 true))))
+          (parse-uuid true))))
   (with-codec {:write-handlers text-write-handlers
                :read-handlers text-read-handlers}
     (check-roundtrip (UUID/fromString "dbd559ef-333b-4f11-96b1-b0654babe844")